Add support for `arc.http` to accept either async or callback functions
