+++ _changelog.md
@@ -10,6 +10,10 @@ Also see: [Architect changelog](https://github.com/architect/architect/blob/main
 
 ### Added
 
+- Added support for `arc.http` to accept either async or callback functions
+  - This is not a breaking change, and `arc.http.async` will remain as a legacy code path
+  - Moving forward we suggest passing all middleware through `arc.http`, which will figure out whether to run in async or callback mode
+  - Note: pass only callback or async middleware, but not both. Don't drink and drive, and don't mix async primitives.
 - Added support for JWE session encryption customization
   - Configure by adding an env var named `ARC_APP_SECRET_ALGO` with a value of `A256GCM` (default if not specified), `A192GCM`, or `A128GCM`
